% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylog.R
\name{mylog}
\alias{mylog}
\title{My logarithm function}
\usage{
mylog(x)
}
\arguments{
\item{x}{is a vector of numerical values that need to be converted into log}
}
\value{
a vector with the same length as \code{x} with logged values of \code{x}
}
\description{
My logarithm function
}
\examples{
log(1)
mylog(1)

}
